remove Task manger flag

+++ streams/src/main/java/org/apache/kafka/streams/processor/internals/assignment/SubscriptionInfo.java
@@ -109,9 +109,9 @@ public class SubscriptionInfo {
 
         if (version >= 8) {
             if (shutdownRequested == null) { //check is required for testing because the mock sets it to null
-                data.setShutdownRequested(0);
+                data.setShutdownRequested((short) 0);
             } else {
-                data.setShutdownRequested(shutdownRequested.get());
+                data.setShutdownRequested((short) shutdownRequested.get());
             }
         }
